2010-06-01 	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libmisc/shell/print-ls.c: Remove (long) cast of st_ino.

+++ cpukit/libmisc/shell/print-ls.c
@@ -132,8 +132,7 @@ printlong(rtems_shell_ls_globals* globals, DISPLAY *dp)
 			continue;
 		sp = p->fts_statp;
 		if (f_inode)
-			(void)printf("%*lu ", dp->s_inode,
-			    (unsigned long)sp->st_ino);
+			(void)printf("%*lu ", dp->s_inode, sp->st_ino);
 		if (f_size && !f_humanize) {
 			(void)printf("%*llu ", dp->s_block,
 			    (long long)howmany(sp->st_blocks, blocksize));
@@ -383,7 +382,7 @@ printaname(rtems_shell_ls_globals* globals,
 	sp = p->fts_statp;
 	chcnt = 0;
 	if (f_inode)
-		chcnt += printf("%*lu ", inodefield, (unsigned long)sp->st_ino);
+		chcnt += printf("%*lu ", inodefield, sp->st_ino);
 	if (f_size) {
 #if RTEMS_REMOVED
 		if (f_humanize) {
